+++ diesel_codegen/src/identifiable.rs
@@ -13,7 +13,7 @@ pub fn derive_identifiable(item: syn::MacroInput) -> Tokens {
         panic!("Could not find a field named `id` on `{}`", &model.name);
     }
 
-    quote!(_Identifiable! {
+    quote!(impl_Identifiable! {
         (
             table_name = #table_name,
             struct_ty = #struct_ty,